LayoutTests/imported/w3c/web-platform-tests/navigation-api/updateCurrentEntry-method/same-document-away-and-back-location-api-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating 'entry1.getState().statevar')

Harness Error (FAIL), message = Unhandled rejection: undefined is not an object (evaluating 'entry1.getState().statevar')

TIMEOUT entry.getState() behavior after navigating away using the location API, then back Test timed out

Harness Error (FAIL), message = Unhandled rejection: undefined is not an object (evaluating 'entry1.getState().statevar')

TIMEOUT entry.getState() behavior after navigating away using the location API, then back Test timed out
